+++ sdk/src/types/mod.rs
@@ -2,13 +2,16 @@ mod json_dag;
 #[cfg(feature = "types")]
 mod nexus_data;
 #[cfg(feature = "types")]
+mod nexus_data_parser;
+#[cfg(feature = "types")]
 mod nexus_objects;
 #[cfg(feature = "types")]
+mod ports_data;
 mod runtime_vertex;
 
 #[cfg(feature = "types")]
 mod serde_parsers;
-#[cfg(feature = "types")]
+mod storage_kind;
 mod tool_meta;
 #[cfg(feature = "types")]
 mod type_name;
@@ -18,10 +21,12 @@ pub use json_dag::*;
 // Only export these for full types feature
 #[cfg(feature = "types")]
 pub use {
-    nexus_data::NexusData,
+    nexus_data::*,
     nexus_objects::NexusObjects,
+    ports_data::PortsData,
     runtime_vertex::RuntimeVertex,
     serde_parsers::*,
+    storage_kind::StorageKind,
     tool_meta::ToolMeta,
     type_name::TypeName,
 };