+++ CHANGELOG.md
@@ -5,23 +5,42 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [`0.3.0`] - Unreleased
+## Unreleased
+
+## [`0.3.0`] - 2025-11-10
 
 ### `nexus-cli`
 
 #### Added
 
-- `nexus crypto init-key --force` wipes the old `crypto` state from config before rotating the key
+- `nexus dag inspect-execution` now also shows the data storage kind for each port
+
+#### Changed
+
+- `nexus dag inspect-execution` now uses new `NexusData` implementation that supports remote storage
+- `nexus dag execute` now uses new `NexusData` implementation that supports remote storage
+
+#### Fixed
+
+- `nexus crypto init-key --force` wipes the old `crypto` state from config before rotating the key to avoid parsing errors
 
 ### `nexus-sdk`
+
+#### Added
+
+- `.nightly-version` file that specifies the Rust nightly version to use
+- `nexus_sdk::nexus` module that holds `NexusClient` functionality to interact with the Nexus network
+- `NexusEventKind::name` method that returns a string representation of the event kind
 
 #### Changed
 
 - standardized array and single value serialization of `NexusData` in `serde_parsers`
+- `NexusData` can now represent data stored remotely in Walrus
 
-#### Added
+#### Fixed
 
-- .nightly-version
+- made faucet requests compatible with old and latest versions of the `sui-faucet`
+- allow skipping the first encrypted message in a new `dh` chain
 
 ## [`0.2.0`] - 2025-08-12